server/pyproject.toml
[project]
name = "tcr-agent-server"
version = "0.1.0"
description = "TCR Dark Matter Explorer — FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "anthropic>=0.49",
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=16",
    "scikit-learn>=1.4",
    "sse-starlette>=2.1",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "duckdb>=1.4.4",
    "google-genai>=1.65.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "aiosqlite>=0.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["."]
